[Storage] Split storage/aptosdb/src/lib.rs into multiple files.
