+++ storage/aptosdb/src/pruner/ledger_pruner/event_store_pruner.rs
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{
+    event_store::EventStore,
     pruner::{db_sub_pruner::DBSubPruner, pruner_utils::get_or_initialize_subpruner_progress},
     schema::db_metadata::{DbMetadataKey, DbMetadataSchema, DbMetadataValue},
-    EventStore,
 };
 use anyhow::Result;
 use aptos_logger::info;
