+++ storage/aptosdb/src/state_store/state_merkle_batch_committer.rs
@@ -4,11 +4,10 @@
 //! This file defines the state merkle snapshot committer running in background thread.
 
 use crate::{
-    jellyfish_merkle_node::JellyfishMerkleNodeSchema,
-    metrics::LATEST_SNAPSHOT_VERSION,
+    metrics::{LATEST_SNAPSHOT_VERSION, OTHER_TIMERS_SECONDS},
+    pruner::PrunerManager,
+    schema::{jellyfish_merkle_node::JellyfishMerkleNodeSchema, version_data::VersionDataSchema},
     state_store::{buffered_state::CommitMessage, StateDb},
-    version_data::VersionDataSchema,
-    PrunerManager, OTHER_TIMERS_SECONDS,
 };
 use anyhow::{anyhow, ensure, Result};
 use aptos_crypto::HashValue;
