+++ storage/aptosdb/src/utils/truncation_helper.rs
@@ -4,22 +4,29 @@
 #![allow(dead_code)]
 
 use crate::{
-    db_metadata::{DbMetadataKey, DbMetadataSchema, DbMetadataValue},
+    common::NUM_STATE_SHARDS,
+    event_store::EventStore,
     ledger_db::{LedgerDb, LedgerDbSchemaBatches},
     schema::{
-        epoch_by_version::EpochByVersionSchema, jellyfish_merkle_node::JellyfishMerkleNodeSchema,
-        ledger_info::LedgerInfoSchema, stale_node_index::StaleNodeIndexSchema,
+        db_metadata::{DbMetadataKey, DbMetadataSchema, DbMetadataValue},
+        epoch_by_version::EpochByVersionSchema,
+        jellyfish_merkle_node::JellyfishMerkleNodeSchema,
+        ledger_info::LedgerInfoSchema,
+        stale_node_index::StaleNodeIndexSchema,
         stale_node_index_cross_epoch::StaleNodeIndexCrossEpochSchema,
-        stale_state_value_index::StaleStateValueIndexSchema, state_value::StateValueSchema,
-        transaction::TransactionSchema, transaction_accumulator::TransactionAccumulatorSchema,
-        transaction_info::TransactionInfoSchema, version_data::VersionDataSchema,
+        stale_state_value_index::StaleStateValueIndexSchema,
+        state_value::StateValueSchema,
+        transaction::TransactionSchema,
+        transaction_accumulator::TransactionAccumulatorSchema,
+        transaction_info::TransactionInfoSchema,
+        version_data::VersionDataSchema,
         write_set::WriteSetSchema,
     },
     state_kv_db::StateKvDb,
     state_merkle_db::StateMerkleDb,
     state_store::MAX_COMMIT_PROGRESS_DIFFERENCE,
+    transaction_store::TransactionStore,
     utils::get_progress,
-    EventStore, TransactionStore, NUM_STATE_SHARDS,
 };
 use anyhow::Result;
 use aptos_jellyfish_merkle::{node_type::NodeKey, StaleNodeIndex};
