+++ storage/backup/backup-cli/src/utils/mod.rs
@@ -19,11 +19,12 @@ use aptos_config::config::{
 use aptos_crypto::HashValue;
 use aptos_db::{
     backup::restore_handler::RestoreHandler,
+    db::AptosDB,
+    get_restore_handler::GetRestoreHandler,
     state_restore::{
         StateSnapshotProgress, StateSnapshotRestore, StateSnapshotRestoreMode, StateValueBatch,
         StateValueWriter,
     },
-    AptosDB, GetRestoreHandler,
 };
 use aptos_infallible::duration_since_epoch;
 use aptos_jellyfish_merkle::{NodeBatch, TreeWriter};
