+++ storage/db-tool/src/replay_verify.rs
@@ -12,7 +12,7 @@ use aptos_config::config::{
     StorageDirPaths, BUFFERED_STATE_TARGET_ITEMS, DEFAULT_MAX_NUM_NODES_PER_LRU_CACHE_SHARD,
     NO_OP_STORAGE_PRUNER_CONFIG,
 };
-use aptos_db::{AptosDB, GetRestoreHandler};
+use aptos_db::{get_restore_handler::GetRestoreHandler, AptosDB};
 use aptos_executor_types::VerifyExecutionMode;
 use aptos_logger::info;
 use aptos_types::transaction::Version;
